Migrate legacy rl: rate-limit keys to canonical rate: namespace
