+++ api/_utils/auth/index.ts
@@ -35,8 +35,6 @@ export {
   PASSWORD_MIN_LENGTH,
   PASSWORD_MAX_LENGTH,
   PASSWORD_BCRYPT_ROUNDS,
-  RATE_LIMIT_PREFIX,
-  RATE_LIMIT_BLOCK_PREFIX,
   RATE_LIMIT_WINDOW_SECONDS,
   RATE_LIMIT_ATTEMPTS,
   CREATE_USER_BLOCK_TTL_SECONDS,
